requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.2
databricks-sdk==0.61.0
datasets==4.0.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
eval_type_backport==0.2.2
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hdf5plugin==5.1.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
intel-cmplr-lib-ur==2025.2.0
intel-openmp==2025.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.15.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mkl==2025.2.0
mkl-service==2.5.2
mkl_fft==2.0.0
mkl_random==1.2.11
mlflow==3.1.4
mlflow-skinny==3.1.4
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numcodecs==0.16.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.98.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentencepiece==0.2.0
sentry-sdk==2.34.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tbb==2022.2.0
tcmlib==1.4.0
threadpoolctl==3.6.0
together==1.5.21
tokenizers==0.19.1
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.40.0
triton==3.3.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umf==0.11.0
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
